% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{to_uppercase}
\alias{to_uppercase}
\title{Convert to uppercase}
\usage{
to_uppercase(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A character string in uppercase.
}
\description{
Converts a string to uppercase using stringr.
}
\examples{
to_uppercase("hello")
}
