% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_a_and_b}
\alias{add_a_and_b}
\title{Add two numbers}
\usage{
add_a_and_b(a, b)
}
\arguments{
\item{a}{A numeric value.}

\item{b}{A numeric value.}
}
\value{
The sum of \code{a} and \code{b}.
}
\description{
This function returns the sum of two numbers.
}
\examples{
add_a_and_b(3, 5)
}
