% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{multiply_x_and_y}
\alias{multiply_x_and_y}
\title{Multiply two numbers}
\usage{
multiply_x_and_y(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The product of \code{x} and \code{y}.
}
\description{
This function multiplies two numbers together.
}
\examples{
multiply_x_and_y(2, 4)
}
